app/requirements.dev.txt
psycopg2-binary

pytest
pytest-env
pytest-cov
flake8-isort
